O banco de dados tava todo bugado, precisei acertar tudo

+++ database/migrations/2025_12_11_015635_produtos.php
@@ -15,25 +15,15 @@ return new class extends Migration
             $table->id();
             $table->string('nome');
             $table->text('descricao');
-            $table->double('preco', 10, 2);
+            $table->decimal('preco', 10, 2);
             $table->string('slug');
             $table->string('imagem')->nullable();
-
-            // Relacionamento com users
             $table->unsignedBigInteger('id_user');
-            $table->foreign('id_user')
-                ->references('id')
-                ->on('users')
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
-
-            // Relacionamento com categorias
+            $table->foreign('id_user')->references('id')->on('users')->onDelete('cascade')->onUpdate('cascade');
+            
             $table->unsignedBigInteger('id_categoria');
-            $table->foreign('id_categoria')
-                ->references('id')
-                ->on('categorias') // <-- aqui estava errado no seu
-                ->onDelete('cascade')
-                ->onUpdate('cascade');
+            $table->foreign('id_categoria')->references('id')->on('categorias')->onDelete('cascade')->onUpdate('cascade');
+
 
             $table->timestamps();
         });
